move stac stuff to separate file
